bot/requirements.txt
python-telegram-bot==21.10
aiohttp==3.11.11
apscheduler==3.10.4
